This commit reverted all demo merge errors/mistakes made.
